Separate parameter related previous session from Prefs
